Dissolve JsonPrinter into the Json.Ast Showable instance
The `JsonPrinter` object was a second, public way to render JSON,
duplicating what the `Json.Ast is Showable` typeclass already exposes via
`.show`. The serialization fold now lives inside that instance, so the
typeclass is the single route to JSON text; `Json is Encodable in Text`
composes with it (compact formatting), and the unused streaming `emit`
variant is dropped. Benchmark and parser-test call sites move to `.show`.
